--- packages/react-native-nitro-modules/cpp/views/ViewComponentDescriptor.cpp
@@ -1,11 +1,0 @@
-//
-// Created by Marc Rousavy on 19.08.26.
-//
-
-#include "ViewComponentDescriptor.hpp"
-
-namespace margelo::nitro {
-
-using namespace facebook;
-
-} // namespace margelo::nitro
