--- packages/react-native-nitro-modules/cpp/views/ViewPropsHolderState.cpp
@@ -1,11 +1,0 @@
-//
-// Created by Marc Rousavy on 19.08.26.
-//
-
-#include "ViewPropsHolderState.hpp"
-
-namespace margelo::nitro {
-
-using namespace facebook;
-
-} // namespace margelo::nitro
