+++ packages/react-native-nitro-modules/cpp/views/ViewPropsHolderState.hpp
@@ -10,4 +10,40 @@ namespace margelo::nitro {
 
 using namespace facebook;
 
+/**
+ * A class that conforms to `react::StateData`, which
+ * only holds the given `TProps` via a `shared_ptr`.
+ *
+ * This is useful to set props via a custom route through
+ * JNI on Android, instead of going through react-native's
+ * JNI/folly prop parser which is type-limited.
+ */
+template <typename TProps>
+struct ViewPropsHolderState final {
+public:
+  ViewPropsHolderState() = default;
+  explicit ViewPropsHolderState(const std::shared_ptr<TProps>& props):
+    _props(props) {}
+
+public:
+  [[nodiscard]]
+  const std::shared_ptr<TProps>& getProps() const {
+    return _props;
+  }
+
+public:
+  #ifdef ANDROID
+  ViewPropsHolderState(const ViewPropsHolderState& /* previousState */, folly::dynamic /* data */) {}
+  folly::dynamic getDynamic() const {
+    throw std::runtime_error("HybridTestViewState does not support folly!");
+  }
+  react::MapBuffer getMapBuffer() const {
+    throw std::runtime_error("HybridTestViewState does not support MapBuffer!");
+  };
+  #endif
+
+private:
+  std::shared_ptr<TProps> _props;
+};
+
 } // namespace margelo::nitro
